+++ e2e/runner/accounting.go
@@ -34,7 +34,7 @@ type Response struct {
 }
 
 func (r *E2ERunner) CheckZRC20BalanceAndSupply() {
-	r.Logger.Info("Checking ZRC20 Balance vs. Supply")
+	r.Logger.Print("Checking ZRC20 Balance vs. Supply")
 
 	err := r.checkETHTSSBalance()
 	require.NoError(r, err, "ETH balance check failed")
@@ -42,8 +42,7 @@ func (r *E2ERunner) CheckZRC20BalanceAndSupply() {
 	err = r.checkERC20TSSBalance()
 	require.NoError(r, err, "ERC20 balance check failed")
 
-	err = r.checkZetaTSSBalance()
-	require.NoError(r, err, "ZETA balance check failed")
+	r.checkZetaTSSBalance()
 
 	err = r.CheckBTCTSSBalance()
 	require.NoError(r, err, "BTC balance check failed")
@@ -181,36 +180,33 @@ func (r *E2ERunner) checkERC20TSSBalance() error {
 		return err
 	}
 	if custodyBalance.Cmp(erc20zrc20Supply) < 0 {
-		return fmt.Errorf("ERC20: TSS balance (%d) < ZRC20 TotalSupply (%d) ", custodyBalance, erc20zrc20Supply)
+		return fmt.Errorf("ERC20: custody balance (%d) < ZRC20 TotalSupply (%d) ", custodyBalance, erc20zrc20Supply)
 	}
 	r.Logger.Info("ERC20: TSS balance (%d) >= ERC20 ZRC20 TotalSupply (%d)", custodyBalance, erc20zrc20Supply)
 	return nil
 }
 
-func (r *E2ERunner) checkZetaTSSBalance() error {
-	zetaLocked, err := r.ConnectorEth.GetLockedAmount(&bind.CallOpts{})
-	if err != nil {
-		return err
-	}
+func (r *E2ERunner) checkZetaTSSBalance() {
+	zetaLockedLegacyConnector, err := r.ZetaEth.BalanceOf(&bind.CallOpts{}, r.ConnectorEthAddr)
+	require.NoError(r, err, "BalanceOf failed for legacy connector")
+
+	zetaLockedConnectorNative, err := r.ZetaEth.BalanceOf(&bind.CallOpts{}, r.ConnectorNativeAddr)
+	require.NoError(r, err, "BalanceOf failed for new connector")
+
+	zetaLocked := big.NewInt(0).Add(zetaLockedLegacyConnector, zetaLockedConnectorNative)
+
 	resp, err := http.Get("http://zetacore0:1317/cosmos/bank/v1beta1/supply/by_denom?denom=azeta")
-	if err != nil {
-		return err
-	}
+	require.NoError(r, err)
 	defer resp.Body.Close()
 	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
+	require.NoError(r, err)
 	var result Response
 	err = json.Unmarshal(body, &result)
-	if err != nil {
-		return err
-	}
+	require.NoError(r, err)
 	zetaSupply, _ := big.NewInt(0).SetString(result.Amount.Amount, 10)
 	if zetaLocked.Cmp(zetaSupply) < 0 {
 		r.Logger.Info("ZETA: TSS balance (%d) < ZRC20 TotalSupply (%d)", zetaLocked, zetaSupply)
 	} else {
 		r.Logger.Info("ZETA: TSS balance (%d) >= ZRC20 TotalSupply (%d)", zetaLocked, zetaSupply)
 	}
-	return nil
 }