+++ testutil/keeper/mocks/crosschain/fungible.go
@@ -420,6 +420,36 @@ func (_m *CrosschainFungibleKeeper) GetUniswapV2Router02Address(ctx types.Contex
 	return r0, r1
 }
 
+// LegacyZETADepositAndCallContract provides a mock function with given fields: ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes
+func (_m *CrosschainFungibleKeeper) LegacyZETADepositAndCallContract(ctx types.Context, sender common.Address, to common.Address, inboundSenderChainID int64, inboundAmount *big.Int, data []byte, indexBytes [32]byte) (*evmtypes.MsgEthereumTxResponse, error) {
+	ret := _m.Called(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
+
+	if len(ret) == 0 {
+		panic("no return value specified for LegacyZETADepositAndCallContract")
+	}
+
+	var r0 *evmtypes.MsgEthereumTxResponse
+	var r1 error
+	if rf, ok := ret.Get(0).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) (*evmtypes.MsgEthereumTxResponse, error)); ok {
+		return rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
+	}
+	if rf, ok := ret.Get(0).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) *evmtypes.MsgEthereumTxResponse); ok {
+		r0 = rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*evmtypes.MsgEthereumTxResponse)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) error); ok {
+		r1 = rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // ProcessAbort provides a mock function with given fields: ctx, inboundSender, amount, outgoing, chainID, coinType, asset, abortAddress, revertMessage
 func (_m *CrosschainFungibleKeeper) ProcessAbort(ctx types.Context, inboundSender string, amount *big.Int, outgoing bool, chainID int64, coinType coin.CoinType, asset string, abortAddress common.Address, revertMessage []byte) (*evmtypes.MsgEthereumTxResponse, error) {
 	ret := _m.Called(ctx, inboundSender, amount, outgoing, chainID, coinType, asset, abortAddress, revertMessage)
@@ -679,36 +709,6 @@ func (_m *CrosschainFungibleKeeper) WithdrawFromGasStabilityPool(ctx types.Conte
 	}
 
 	return r0
-}
-
-// ZETADepositAndCallContract provides a mock function with given fields: ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes
-func (_m *CrosschainFungibleKeeper) ZETADepositAndCallContract(ctx types.Context, sender common.Address, to common.Address, inboundSenderChainID int64, inboundAmount *big.Int, data []byte, indexBytes [32]byte) (*evmtypes.MsgEthereumTxResponse, error) {
-	ret := _m.Called(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
-
-	if len(ret) == 0 {
-		panic("no return value specified for ZETADepositAndCallContract")
-	}
-
-	var r0 *evmtypes.MsgEthereumTxResponse
-	var r1 error
-	if rf, ok := ret.Get(0).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) (*evmtypes.MsgEthereumTxResponse, error)); ok {
-		return rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
-	}
-	if rf, ok := ret.Get(0).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) *evmtypes.MsgEthereumTxResponse); ok {
-		r0 = rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
-	} else {
-		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(*evmtypes.MsgEthereumTxResponse)
-		}
-	}
-
-	if rf, ok := ret.Get(1).(func(types.Context, common.Address, common.Address, int64, *big.Int, []byte, [32]byte) error); ok {
-		r1 = rf(ctx, sender, to, inboundSenderChainID, inboundAmount, data, indexBytes)
-	} else {
-		r1 = ret.Error(1)
-	}
-
-	return r0, r1
 }
 
 // ZETARevertAndCallContract provides a mock function with given fields: ctx, sender, to, inboundSenderChainID, destinationChainID, remainingAmount, data, indexBytes