+++ changelog.md
@@ -8,6 +8,7 @@
 * [3958](https://github.com/zeta-chain/node/pull/3958) - integrate TON http-rpc
 * [3961](https://github.com/zeta-chain/node/pull/3961) - cosmos v53 upgrade
 * [3977](https://github.com/zeta-chain/node/pull/3977) - integrate new ton features (call, increase_seqno, compliance)
+* [3990](https://github.com/zeta-chain/node/pull/3990) - add support for deposits using the new v2 connector contract
 
 ### Fixes
 
@@ -30,6 +31,7 @@
 * [3975](https://github.com/zeta-chain/node/pull/3975) - add v2 connector deployment to the e2e test.
 * [3976](https://github.com/zeta-chain/node/pull/3976) - add connector fund migration e2e test using contracts only.
 * [3999](https://github.com/zeta-chain/node/pull/3999) - run simulation tests nightly
+* [3985](https://github.com/zeta-chain/node/pull/3985)- add e2e tests for deposit and withdraw with big payload
 
 ## v31.0.0